requirements.txt
Django>=1.8,<2.0
django-heroku==0.2.0
djangorestframework==3.7.7
gunicorn==19.7.1
psycopg2-binary==2.7.4
pycodestyle==2.3.1
Pygments==2.2.0
